+++ Game/include/Entity/EntityCoin.hpp
@@ -14,7 +14,7 @@ class EntityCoin : public Entity {
   ~EntityCoin() override = default;
 
   void onInit() override;
-  void onUpdate(double delta) override;
+  void onUpdate(float delta) override;
   void onRender() override;
   void onCollision(std::shared_ptr<IEntity> collision_entity) override;
