+++ Game/include/Entity/EntityGround.hpp
@@ -12,7 +12,7 @@ class EntityGround : public IEntity {
   ~EntityGround() override = default;
 
   void onInit() override;
-  void onUpdate(double delta) override;
+  void onUpdate(float delta) override;
   void onRender() override;
   void onCollision(std::shared_ptr<IEntity> collision_entity) override;
