+++ Game/include/Entity/IEntity.hpp
@@ -14,7 +14,7 @@ class IEntity {
   virtual ~IEntity() = default;
 
   virtual void onInit() = 0;
-  virtual void onUpdate(double delta) = 0;
+  virtual void onUpdate(float delta) = 0;
   virtual void onRender() = 0;
   virtual void onCollision(std::shared_ptr<IEntity> collision_entity) = 0;
